fix: stake token label
